--- Signlanguage/app/src/main/java/com/example/singlanguage/Fragment1_2.java
@@ -1,27 +1,0 @@
-package com.example.singlanguage;
-
-import android.content.Context;
-import android.net.Uri;
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-
-public class Fragment1_2 extends Fragment {
-
-    public Fragment1_2() {
-        // Required empty public constructor
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_fragment1_2, container, false);
-    }
-}
-
