+++ Signlanguage/app/src/main/java/com/example/singlanguage/Fragment4.java
@@ -1,0 +1,25 @@
+package com.example.singlanguage;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+
+public class Fragment4 extends Fragment {
+
+    public Fragment4() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_fragment4, container, false);
+    }
+}
+
