+++ Signlanguage/app/src/main/java/com/example/singlanguage/MyPagerAdapter.java
@@ -21,14 +21,23 @@ public class MyPagerAdapter extends FragmentPagerAdapter {
 
         switch (position) {
             case 0:
-                Fragment1 tab1 = new Fragment1();
+                Fragment1 tab1 = new Fragment1();   //숫자
                 return tab1;
             case 1:
-                Fragment2 tab2 = new Fragment2();
+                Fragment2 tab2 = new Fragment2();   //자모음
                 return tab2;
             case 2:
-                Fragment3 tab3 = new Fragment3();
+                Fragment3 tab3 = new Fragment3();   //사물
                 return tab3;
+            case 3:
+                Fragment4 tab4 = new Fragment4();   //사람
+                return tab4;
+            case 4:
+                Fragment5 tab5 = new Fragment5();   //음식
+                return tab5;
+            case 5:
+                Fragment6 tab6 = new Fragment6();   //기타
+                return tab6;
             default:
                 return null;
         }
