+++ Signlanguage/app/src/main/res/layout/activity_category.xml
@@ -13,10 +13,8 @@
         android:id="@+id/textView"
         android:layout_width="match_parent"
         android:layout_height="86dp"
-        android:layout_centerHorizontal="true"
         android:text="암기할 단어를 선택해주세요."
-        android:translationX="100dp"
-        android:translationY="30dp"
+        android:gravity="center"
         app:layout_constraintBottom_toTopOf="@+id/tabs"
         app:layout_constraintHorizontal_bias="0.135"
         app:layout_constraintLeft_toLeftOf="parent"
