+++ Signlanguage/app/src/main/res/layout/fragment_fragment3.xml
@@ -1,14 +1,187 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:orientation="vertical"
     tools:context=".Fragment3">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="~형용사 수화들~" />
+        android:layout_height="match_parent">
 
-</FrameLayout>
+        <GridLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:columnCount="4"
+            android:layout_gravity="center_horizontal">
+
+            <LinearLayout
+                android:id="@+id/layout_zero"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_zero"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/zero" />
+
+                <TextView
+                    android:id="@+id/txt_zero"
+                    style="@style/CustomTextView"
+                    android:text="0" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_one"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_one"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/btn_clk_one" />
+
+                <TextView
+                    android:id="@+id/txt_one"
+                    style="@style/CustomTextView"
+                    android:text="1" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_two"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_two"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/btn_clk_two" />
+
+                <TextView
+                    android:id="@+id/txt_two"
+                    style="@style/CustomTextView"
+                    android:text="2" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_three"
+                style="@style/CustomLinearLayout_ver">
+                <ImageView
+                    android:id="@+id/image_three"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/three" />
+
+                <TextView
+                    android:id="@+id/txt_three"
+                    style="@style/CustomTextView"
+                    android:text="3" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_four"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_four"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/four" />
+
+                <TextView
+                    android:id="@+id/txt_four"
+                    style="@style/CustomTextView"
+                    android:text="4" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_five"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_five"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/five" />
+
+                <TextView
+                    android:id="@+id/txt_five"
+                    style="@style/CustomTextView"
+                    android:text="5" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_six"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_six"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/six" />
+
+                <TextView
+                    android:id="@+id/txt_six"
+                    style="@style/CustomTextView"
+                    android:text="6" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_seven"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_seven"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/seven" />
+
+                <TextView
+                    android:id="@+id/txt_seven"
+                    style="@style/CustomTextView"
+                    android:text="7" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_eight"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_eight"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/eight" />
+
+                <TextView
+                    android:id="@+id/txt_eight"
+                    style="@style/CustomTextView"
+                    android:text="8" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_nine"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_nine"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/nine" />
+
+                <TextView
+                    android:id="@+id/txt_nine"
+                    style="@style/CustomTextView"
+                    android:text="9" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/layout_ten"
+                style="@style/CustomLinearLayout_ver">
+
+                <ImageView
+                    android:id="@+id/image_ten"
+                    style="@style/CustomImageView"
+                    android:src="@drawable/ten" />
+
+                <TextView
+                    android:id="@+id/txt_ten"
+                    style="@style/CustomTextView"
+                    android:text="10" />
+            </LinearLayout>
+        </GridLayout>
+    </ScrollView>
+
+</LinearLayout>
