+++ Signlanguage/app/src/main/res/raw/capdb.csv
@@ -1,17 +1,17 @@
-﻿기역,자음,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-니은,자음,nieun,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-디귿,자음,digeut,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-리을,자음,rieul,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-미음,자음,mieum,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-비읍,자음,bieup,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-시옷,자음,siot,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-이응,자음,ieung,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-지읒,자음,jieut,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-치읓,자음,chieut,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-키읔,자음,kieuk,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-티읕,자음,tieut,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-피읖,자음,pieup,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-히읗,자음,hieut,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
+﻿기역,자음,giyeok,오른 주먹의 1·5지를 펴서 1지 끝이 아래로 손등이 밖으로 향하게 세운다.,ㄱ(기역) [명사]한글 자모의 첫째 글자.
+니은,자음,nieun,오른 주먹의 1·5지를 펴서 1지 끝이 왼쪽으로 손등이 밖으로 향하게 한다.,ㄴ(니은) [명사] 한글 자모의 둘째 글자.
+디귿,자음,digeut,오른 주먹의 1·2지를 펴서 벌려 끝이 왼쪽으로 손등이 밖으로 향하게 한다.,ㄷ(디귿) [명사]한글 자모의 셋째 글자.
+리을,자음,rieul,오른 주먹의 1·2·3지를 펴서 벌려 끝이 왼쪽으로 손등이 밖으로 향하게 한다.,ㄹ(리을) [명사]한글 자모의 넷째 글자.
+미음,자음,mieum,손등이 안으로 향하게 세워 쥔 오른 주먹의 1·2지를 구부린다.,ㅁ(미음) [명사]한글 자모의 다섯째 글자.
+비읍,자음,bieup,오른 주먹의 1·2·3·4지를 펴서 붙여 손바닥이 밖으로 손끝이 위로 향하게 세운다.,ㅂ(비읍) [명사]한글 자모의 여섯째 글자.
+시옷,자음,siot,오른 주먹의 1·2지를 펴서 벌려 등이 밖으로 끝이 아래로 향하게 한다.,ㅅ(시옷) [명사] 한글 자모의 일곱째 글자.
+이응,자음,ieung,손바닥이 밖으로 향하게 펴서 세운 오른손의 1·5지 끝을 맞대어 동그라미를 만들어 보인다.,ㅇ(이응) [명사] 한글 자모의 여덟째 글자.
+지읒,자음,jieut,오른 주먹의 1·2·5지를 펴서 등이 밖으로 끝이 아래로 향하게 세운다.,ㅈ(지읒) [명사]한글 자모의 아홉째 글자.
+치읓,자음,chieut,오른 주먹의 1·2·3·5지를 펴서 등이 밖으로 끝이 아래로 향하게 세운다.,ㅊ(치읓) [명사]한글 자모의 열째 글자.
+키읔,자음,kieuk,"오른 주먹의 1지를 반쯤 굽히고 2·5지를 펴서 등이 밖으로, 끝이 아래로 향하게 세운다.",ㅋ(키읔) [명사]한글 자모의 열한째 글자.
+티읕,자음,tieut,손등이 밖으로 향하게 쥔 오른 주먹의 1·2·3지를 펴서 끝이 왼쪽으로 향하게 하여 2·3지는 붙이고 1지는 뗀다.,ㅌ(티읕) [명사]한글 자모의 열두째 글자.
+피읖,자음,pieup,손등이 안으로 향하게 세운 오른손의 1·2·3·4·5지를 반쯤 굽힌다.,ㅍ(피읖) [명사]한글 자모의 열셋째 글자.
+히읗,자음,hieut,오른 주먹의 5지를 펴서 바닥이 밖으로 향하게 세운다.,ㅎ(히읗) [명사] 한글 자모의 열넷째 글자.
 아,모음,a,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
 야,모음,ya,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
 어,모음,eo,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
@@ -35,7 +35,7 @@
 4,숫자,four,5지를 접고 1·2·3·4지를 펴서 벌려 바닥이 밖으로 향하게 세운다.,사 (四) [수사] 삼에 일을 더한 수. 아라비아 숫자로는 ‘4’ 로마 숫자로는 ‘Ⅳ’로 쓴다.
 5,숫자,five,손등이 안으로 향하게 쥔 오른 주먹의 5지를 펴서 눕힌다.,오 (五) [수사] 사에 일을 더한 수. 아라비아 숫자로는 ‘5’ 로마 숫자로는 ‘Ⅴ’로 쓴다.
 6,숫자,six,오른 주먹의 1·5지를 펴서 등이 밖으로 향하게 모로 세운다.,육 (六) [수사] 오에 일을 더한 수. 아라비아 숫자로는 ‘6’ 로마 숫자로는 ‘Ⅵ’으로 쓴다.
-7,숫자,seven,오른 주먹의 1·2·5지를 펴서 벌려 등이 밖으로 향하게 모로 세운다.,"칠 (七) [수사] 육에 일을 더한 수. 아라비아 숫자로는 ‘7’, 로마 숫자로는 ‘Ⅶ’로 쓴다."
+7,숫자,seven,오른 주먹의 1·2·5지를 펴서 벌려 등이 밖으로 향하게 모로 세운다.,칠 (七) [수사] 육에 일을 더한 수. 아라비아 숫자로는 ‘7’ 로마 숫자로는 ‘Ⅶ’로 쓴다.
 8,숫자,eight,4지만 접고 나머지 손가락을 펴서 벌린 오른손을 등이 밖으로 끝이 왼쪽으로 향하게 한다.,팔 (八) [수사] 칠에 일을 더한 수. 아라비아 숫자로는 ‘8’ 로마 숫자로는 ‘Ⅷ’로 쓴다.
 9,숫자,nine,오른손을 펴서 손바닥이 안으로 향하게 비스듬히 세운다.,구 (九) [수사] 팔에 일을 더한 수. 아라비아 숫자로는 ‘9’ 로마 숫자로는 ‘Ⅸ’로 쓴다.
 10,숫자,ten,오른손의 1·5지 끝을 맞대어 동그라미를 만들어 손바닥이 밖으로 향하게 세운다.,십 (十) [수사] 구에 일을 더한 수. 아라비아 숫자로는 ‘10’ 로마 숫자로는 ‘X’으로 쓴다.
@@ -54,7 +54,7 @@
 여자,사람,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
 집,사물,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
 가루,사물,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-북쪽,방향,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
-남쪽,방향,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
+북쪽,기타,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
+남쪽,기타,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
 상의하다,기타,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
 열매,음식,giyeok,오른 주먹의 1지를 펴서 바닥이 밖으로 향하게 세운다.,하나 [수사] 수효를 세는 맨 처음 수.
